test: ArtDetailQueryRepository pre & post 작업 처리 전용 컴포넌트 분리 (ArtDetailQueryRepositoryTestSupporter)
